refactor(assistant): move auto-analysis bridge next to the analyze-conversation service
