+++ assistant/src/__tests__/auto-analysis-end-to-end.test.ts
@@ -130,7 +130,6 @@ mock.module("../runtime/services/analyze-conversation.js", () => ({
 
 // ── Real imports ──────────────────────────────────────────────────
 
-import { conversationAnalyzeJob } from "../memory/conversation-analyze-job.js";
 import { indexMessageNow } from "../memory/indexer.js";
 import { createConversation } from "../persistence/conversation-crud.js";
 import { getDb, getMemoryDb } from "../persistence/db-connection.js";
@@ -141,6 +140,7 @@ import {
   memoryJobs,
   messages,
 } from "../persistence/schema/index.js";
+import { conversationAnalyzeJob } from "../runtime/services/conversation-analyze-job.js";
 import { setOverridesForTesting } from "./feature-flag-test-helpers.js";
 
 // ── Helpers ───────────────────────────────────────────────────────
