+++ assistant/src/daemon/__tests__/conversation-lifecycle-auto-analyze.test.ts
@@ -66,7 +66,7 @@ mock.module("../../config/loader.js", () => ({
   loadConfig: () => ({ memory: { v2: { enabled: v2Enabled } } }),
 }));
 
-mock.module("../../memory/auto-analysis-guard.js", () => ({
+mock.module("../../runtime/services/auto-analysis-guard.js", () => ({
   AUTO_ANALYSIS_SOURCE: "auto-analysis",
   isAutoAnalysisConversation: (conversationId: string) =>
     autoAnalysisConversations.has(conversationId),
@@ -82,8 +82,8 @@ mock.module("../../persistence/jobs-store.js", () => ({
 }));
 
 const realAutoAnalysisEnqueue =
-  await import("../../memory/auto-analysis-enqueue.js");
-mock.module("../../memory/auto-analysis-enqueue.js", () => ({
+  await import("../../runtime/services/auto-analysis-enqueue.js");
+mock.module("../../runtime/services/auto-analysis-enqueue.js", () => ({
   ...realAutoAnalysisEnqueue,
   enqueueAutoAnalysisIfEnabled: (args: {
     conversationId: string;
@@ -297,7 +297,7 @@ describe("disposeConversation — auto-analysis enqueue", () => {
     // the cleanup chain runs.
     autoAnalyzeEnabled = true;
 
-    mock.module("../../memory/auto-analysis-guard.js", () => ({
+    mock.module("../../runtime/services/auto-analysis-guard.js", () => ({
       AUTO_ANALYSIS_SOURCE: "auto-analysis",
       isAutoAnalysisConversation: () => {
         throw new Error("db closed");
@@ -318,7 +318,7 @@ describe("disposeConversation — auto-analysis enqueue", () => {
     expect(autoAnalyzeCalls).toHaveLength(1);
 
     // Restore the non-throwing stub for subsequent tests.
-    mock.module("../../memory/auto-analysis-guard.js", () => ({
+    mock.module("../../runtime/services/auto-analysis-guard.js", () => ({
       AUTO_ANALYSIS_SOURCE: "auto-analysis",
       isAutoAnalysisConversation: (conversationId: string) =>
         autoAnalysisConversations.has(conversationId),
@@ -334,7 +334,7 @@ describe("disposeConversation — auto-analysis enqueue", () => {
     autoAnalyzeEnabled = true;
 
     // Temporarily re-mock the helper to throw.
-    mock.module("../../memory/auto-analysis-enqueue.js", () => ({
+    mock.module("../../runtime/services/auto-analysis-enqueue.js", () => ({
       enqueueAutoAnalysisIfEnabled: () => {
         throw new Error("boom");
       },
@@ -351,7 +351,7 @@ describe("disposeConversation — auto-analysis enqueue", () => {
     expect(memoryJobCalls).toHaveLength(1);
 
     // Restore the non-throwing stub so other tests aren't affected.
-    mock.module("../../memory/auto-analysis-enqueue.js", () => ({
+    mock.module("../../runtime/services/auto-analysis-enqueue.js", () => ({
       enqueueAutoAnalysisIfEnabled: (args: {
         conversationId: string;
         trigger: "batch" | "idle" | "lifecycle";
