+++ assistant/src/memory/context-search/sources/conversations.ts
@@ -3,11 +3,11 @@ import {
   buildFtsMatchQuery,
   buildRecallEvidenceExcerpt,
 } from "../../../persistence/conversation-queries.js";
+import { AUTO_ANALYSIS_SOURCE } from "../../../runtime/services/auto-analysis-guard.js";
 import {
   parseExternalContentEnvelope,
   wrapUntrustedContent,
 } from "../../../security/untrusted-content.js";
-import { AUTO_ANALYSIS_SOURCE } from "../../auto-analysis-guard.js";
 import { rawAll } from "../../raw-query.js";
 import type { RecallSearchContext, RecallSearchResult } from "../types.js";
