+++ assistant/src/memory/register-job-handlers.ts
@@ -2,8 +2,8 @@ import type { AssistantConfig } from "../config/types.js";
 import type { MemoryJob } from "../persistence/jobs-store.js";
 import { registerJobHandler } from "../persistence/jobs-worker.js";
 import { maintainJob as memoryV3MaintainJob } from "../plugins/defaults/memory/v3/maintain-job.js";
+import { conversationAnalyzeJob } from "../runtime/services/conversation-analyze-job.js";
 import { getLogger } from "../util/logger.js";
-import { conversationAnalyzeJob } from "./conversation-analyze-job.js";
 import { bootstrapFromHistory } from "./graph/bootstrap.js";
 import { runConsolidation } from "./graph/consolidation.js";
 import { runDecayTick } from "./graph/decay.js";
