+++ assistant/src/runtime/services/conversation-analyze-job.ts
@@ -10,10 +10,10 @@
 // logic and recursion guard apply.
 // ---------------------------------------------------------------------------
 
-import type { AssistantConfig } from "../config/types.js";
-import type { MemoryJob } from "../persistence/jobs-store.js";
-import { analyzeConversation } from "../runtime/services/analyze-conversation.js";
-import { getLogger } from "../util/logger.js";
+import type { AssistantConfig } from "../../config/types.js";
+import type { MemoryJob } from "../../persistence/jobs-store.js";
+import { getLogger } from "../../util/logger.js";
+import { analyzeConversation } from "./analyze-conversation.js";
 import { enqueueAutoAnalysisIfEnabled } from "./auto-analysis-enqueue.js";
 
 const log = getLogger("conversation-analyze-job");
